requirements.txt
streamlit==1.32.0
tensorflow==2.16.1
numpy
pandas
matplotlib
seaborn
scikit-learn
pillow